+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/HibernateAccelerableHqlPresetLookupDao.java
@@ -13,6 +13,8 @@ import org.hibernate.query.Query;
 import org.springframework.orm.hibernate5.HibernateTemplate;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -145,47 +147,63 @@ public class HibernateAccelerableHqlPresetLookupDao<E extends Entity<?>, PE exte
     @Override
     public List<E> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         try {
-            if (nativeLookup.supportPreset(preset)) {
-                List<E> result = template.executeWithNativeSession(session -> session.doReturningWork(
-                        connection -> this.nativeLookup.lookupEntity(connection, preset, objs, pagingInfo)
-                ));
-                assert result != null;
-                return result;
-            } else {
-                return template.execute(session -> {
-                    // 构建 HqlCondition 对象。
-                    HqlCondition condition = new HqlCondition(classPE, entityAlias);
-
-                    // 调用 presetConditionMaker 相应方法，补充查询信息。
-                    presetConditionMaker.makeCondition(condition, preset, objs);
-
-                    // 创建 Query 对象。
-                    condition.setQueryType(QueryType.ENTITY);
-                    HqlQueryInfo queryInfo = HqlQueryInfoFactory.buildQueryInfoFormCondition(condition);
-                    // HQL 的安全性由 presetConditionMaker 保证，因此这里不需要进行检查。
-                    @SuppressWarnings("SqlSourceToSinkFlow")
-                    Query<PE> query = session.createQuery(queryInfo.getHql(), classPE);
-
-                    // 设置参数。
-                    for (Map.Entry<String, Object> entry : queryInfo.getParamMap().entrySet()) {
-                        String paramName = entry.getKey();
-                        Object paramValue = entry.getValue();
-                        query.setParameter(paramName, paramValue);
-                    }
-
-                    // 设置分页信息。
-                    query.setFirstResult(pagingInfo.getPage() * pagingInfo.getRows());
-                    query.setMaxResults(pagingInfo.getRows());
-
-                    // 执行查询。
-                    List<PE> pes = query.getResultList();
-
-                    // 将结果转换为相应的实体。
-                    return pes.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());
-                });
+            // 展开参数。
+            int page = pagingInfo.getPage();
+            int rows = pagingInfo.getRows();
+            // 每页行数大于 0 时，按照正常的逻辑查询数据。
+            if (rows > 0) {
+                return lookupWithPositiveRows(preset, objs, pagingInfo, page, rows);
+            }
+            // 否则返回空列表。
+            else {
+                return Collections.emptyList();
             }
         } catch (Exception e) {
             throw new DaoException(e);
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Nullable
+    private List<E> lookupWithPositiveRows(String preset, Object[] objs, PagingInfo pagingInfo, int page, int rows) {
+        if (nativeLookup.supportPreset(preset)) {
+            List<E> result = template.executeWithNativeSession(session -> session.doReturningWork(
+                    connection -> this.nativeLookup.lookupEntity(connection, preset, objs, pagingInfo)
+            ));
+            assert result != null;
+            return result;
+        } else {
+            return template.execute(session -> {
+                // 构建 HqlCondition 对象。
+                HqlCondition condition = new HqlCondition(classPE, entityAlias);
+
+                // 调用 presetConditionMaker 相应方法，补充查询信息。
+                presetConditionMaker.makeCondition(condition, preset, objs);
+
+                // 创建 Query 对象。
+                condition.setQueryType(QueryType.ENTITY);
+                HqlQueryInfo queryInfo = HqlQueryInfoFactory.buildQueryInfoFormCondition(condition);
+                // HQL 的安全性由 presetConditionMaker 保证，因此这里不需要进行检查。
+                @SuppressWarnings("SqlSourceToSinkFlow")
+                Query<PE> query = session.createQuery(queryInfo.getHql(), classPE);
+
+                // 设置参数。
+                for (Map.Entry<String, Object> entry : queryInfo.getParamMap().entrySet()) {
+                    String paramName = entry.getKey();
+                    Object paramValue = entry.getValue();
+                    query.setParameter(paramName, paramValue);
+                }
+
+                // 设置分页信息。
+                query.setFirstResult(page * rows);
+                query.setMaxResults(rows);
+
+                // 执行查询。
+                List<PE> pes = query.getResultList();
+
+                // 将结果转换为相应的实体。
+                return pes.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());
+            });
         }
     }