+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/HibernateAccelerablePresetLookupDao.java
@@ -15,6 +15,7 @@ import org.hibernate.criterion.Projections;
 import org.springframework.orm.hibernate5.HibernateTemplate;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -108,23 +109,39 @@ public class HibernateAccelerablePresetLookupDao<E extends Entity<?>, PE extends
     @Override
     public List<E> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         try {
-            if (nativeLookup.supportPreset(preset)) {
-                List<E> result = template.executeWithNativeSession(session -> session.doReturningWork(
-                        connection -> this.nativeLookup.lookupEntity(connection, preset, objs, pagingInfo)
-                ));
-                assert result != null;
-                return result;
-            } else {
-                DetachedCriteria criteria = DetachedCriteria.forClass(classPE);
-                presetCriteriaMaker.makeCriteria(criteria, preset, objs);
-                @SuppressWarnings("unchecked")
-                List<PE> byCriteria = (List<PE>) template.findByCriteria(
-                        criteria, pagingInfo.getPage() * pagingInfo.getRows(), pagingInfo.getRows()
-                );
-                return byCriteria.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());
+            // 展开参数。
+            int page = pagingInfo.getPage();
+            int rows = pagingInfo.getRows();
+            // 每页行数大于 0 时，按照正常的逻辑查询数据。
+            if (rows > 0) {
+                return lookupWithPositiveRows(preset, objs, pagingInfo, page, rows);
+            }
+            // 否则返回空列表。
+            else {
+                return Collections.emptyList();
             }
         } catch (Exception e) {
             throw new DaoException(e);
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Nonnull
+    private List<E> lookupWithPositiveRows(String preset, Object[] objs, PagingInfo pagingInfo, int page, int rows) {
+        if (nativeLookup.supportPreset(preset)) {
+            List<E> result = template.executeWithNativeSession(session -> session.doReturningWork(
+                    connection -> this.nativeLookup.lookupEntity(connection, preset, objs, pagingInfo)
+            ));
+            assert result != null;
+            return result;
+        } else {
+            DetachedCriteria criteria = DetachedCriteria.forClass(classPE);
+            presetCriteriaMaker.makeCriteria(criteria, preset, objs);
+            @SuppressWarnings("unchecked")
+            List<PE> byCriteria = (List<PE>) template.findByCriteria(
+                    criteria, page * rows, rows
+            );
+            return byCriteria.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());
         }
     }