+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/JdbcPresetLookupDao.java
@@ -10,6 +10,7 @@ import org.springframework.jdbc.core.ArgumentPreparedStatementSetter;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -50,12 +51,21 @@ public class JdbcPresetLookupDao<E extends Entity<?>> implements PresetLookupDao
     @Override
     public List<E> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         try {
-            SQLAndParameter sqlAndParameter = processor.providePresetPaging(preset, objs, pagingInfo);
-            return template.query(
-                    sqlAndParameter.getSql(),
-                    new ArgumentPreparedStatementSetter(sqlAndParameter.getFirstParameters()),
-                    processor::resolvePresetPaging
-            );
+            // 展开参数。
+            int rows = pagingInfo.getRows();
+            // 每页行数大于 0 时，按照正常的逻辑查询数据。
+            if (rows > 0) {
+                SQLAndParameter sqlAndParameter = processor.providePresetPaging(preset, objs, pagingInfo);
+                return template.query(
+                        sqlAndParameter.getSql(),
+                        new ArgumentPreparedStatementSetter(sqlAndParameter.getFirstParameters()),
+                        processor::resolvePresetPaging
+                );
+            }
+            // 否则返回空列表。
+            else {
+                return Collections.emptyList();
+            }
         } catch (Exception e) {
             throw new DaoException(e);
         }