+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/MybatisPresetLookupDao.java
@@ -8,6 +8,7 @@ import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.mybatis.spring.SqlSessionTemplate;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -58,7 +59,16 @@ public class MybatisPresetLookupDao<E extends Entity<?>> implements PresetLookup
     @Override
     public List<E> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         try {
-            return template.selectList(concatId(presetPagingId), new PresetLookupInfo(preset, objs, pagingInfo));
+            // 展开参数。
+            int rows = pagingInfo.getRows();
+            // 每页行数大于 0 时，按照正常的逻辑查询数据。
+            if (rows > 0) {
+                return template.selectList(concatId(presetPagingId), new PresetLookupInfo(preset, objs, pagingInfo));
+            }
+            // 否则返回空列表。
+            else {
+                return Collections.emptyList();
+            }
         } catch (Exception e) {
             throw new DaoException(e);
         }