+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/RedisEntireLookupDao.java
@@ -11,6 +11,7 @@ import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.springframework.data.redis.core.RedisTemplate;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -67,13 +68,28 @@ public class RedisEntireLookupDao<K extends Key, E extends Entity<K>, JE extends
     @Override
     public List<E> lookup(PagingInfo pagingInfo) throws DaoException {
         try {
-            int beginIndex = pagingInfo.getPage() * pagingInfo.getRows();
-            int endIndex = beginIndex + pagingInfo.getRows();
-            List<E> es = internalLookup();
-            return es.subList(beginIndex, Math.min(es.size(), endIndex));
+            // 展开参数。
+            int page = pagingInfo.getPage();
+            int rows = pagingInfo.getRows();
+            // 每页行数大于 0 时，按照正常的逻辑查询数据。
+            if (rows > 0) {
+                return lookupWithPositiveRows(page, rows);
+            }
+            // 否则返回空列表。
+            else {
+                return Collections.emptyList();
+            }
         } catch (Exception e) {
             throw new DaoException(e);
         }
+    }
+
+    @Nonnull
+    private List<E> lookupWithPositiveRows(int page, int rows) {
+        int beginIndex = page * rows;
+        int endIndex = beginIndex + rows;
+        List<E> es = internalLookup();
+        return es.subList(beginIndex, Math.min(es.size(), endIndex));
     }
 
     @Override