+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/bean/dto/PagingUtil.java
@@ -13,6 +13,21 @@ import java.util.stream.Collectors;
 /**
  * 分页工具类。
  *
+ * <p>
+ * 需要注意的是，在 1.5.0 版本中，查询服务、缓存、数据访问层应用了新的分页特性，
+ * 即当分页信息的页数小于 0 或每页行数小于 0 时，查询全部数据。<br>
+ * 上述新的特性只应用在了服务、缓存、数据访问层中，而没有应用在工具类中。也就是说，您在调用工具类中的方法时，
+ * 仍然需要保证分页信息的页数和行数是合法的。<br>
+ *
+ * <p>
+ * 在 1.5.0 版本之前，该工具类默认输入数据合法，即页数大于等于 0, 行数大于等于 0。但在大量的实践中发现，
+ * 部分开发人员习惯使用页数或每页行数小于 0 的分页信息
+ * （即 {@link PagingInfo#getPage()} 为负数或 {@link PagingInfo#getRows()} 为负数）表示不分页。
+ * 这些行为导致工具类经常以预期之外的方式工作，并产生了一些不可预期的结果。<br>
+ * 在 1.5.0 版本中，该工具类不再默认输入数据合法，并且对输入数据的合法性进行了检查，在输入数据不合法时，
+ * 会抛出异常。这样做的目的是为了让工具类的行为更加可预测，更加符合预期。<br>
+ * 应用合法性校验的方法会在方法的文档中进行标注。在 1.5.0 版本之前已经弃用的方法保持原有的行为，不会进行合法性校验。
+ *
  * @author DwArFeng
  * @since 0.0.3-beta
  */
@@ -25,20 +40,38 @@ public final class PagingUtil {
     /**
      * 根据分页信息、数据总量、当前页的数据构造一个分页信息。
      *
+     * <p>
+     * 必须保证 <code>pagingInfo</code> 参数的页数和行数是合法的，即
+     * <code>page &gt;= 0</code> 且 <code>rows &gt;= 0</code>，否则会抛出异常。
+     *
+     * <p>
+     * 必须保证 <code>count</code> 参数是合法的，即 <code>count &gt;= 0</code>，否则会抛出异常。
+     *
      * @param pagingInfo 分页信息。
      * @param count      总数据量。
      * @param data       当前页的数据。
      * @param <E>        数据的类型。
      * @return 构造的分页信息。
+     * @throws IllegalArgumentException 参数不合法。
      */
     public static <E> PagedData<E> pagedData(PagingInfo pagingInfo, int count, List<E> data) {
-        return new PagedData<>(
-                pagingInfo.getPage(),
-                count <= 0 || pagingInfo.getRows() <= 0 ? 0 : (count - 1) / pagingInfo.getRows() + 1,
-                pagingInfo.getRows(),
-                count,
-                data
-        );
+        int page = pagingInfo.getPage();
+        int rows = pagingInfo.getRows();
+        if (page < 0 || rows < 0) {
+            throw new IllegalArgumentException("分页信息的页数必须大于等于 0, 行数必须大于等于 0");
+        }
+        if (count < 0) {
+            throw new IllegalArgumentException("数据总量必须大于等于 0");
+        }
+        int totalPages;
+        if (rows == 0) {
+            totalPages = -1;
+        } else if (count == 0) {
+            totalPages = 0;
+        } else {
+            totalPages = (count - 1) / rows + 1;
+        }
+        return new PagedData<>(pagingInfo.getPage(), totalPages, rows, count, data);
     }
 
     /**
@@ -49,13 +82,7 @@ public final class PagingUtil {
      * @return 构造的分页信息。
      */
     public static <E> PagedData<E> pagedData(List<E> data) {
-        return new PagedData<>(
-                0,
-                1,
-                data.size(),
-                data.size(),
-                data
-        );
+        return new PagedData<>(0, 1, data.size(), data.size(), data);
     }
 
     /**
@@ -184,8 +211,9 @@ public final class PagingUtil {
      * @return 第二个类型的PagedData。
      */
     @SuppressWarnings("DuplicatedCode")
-    public static <U extends Bean, V extends Bean> PagedData<V>
-    transform(@Nonnull PagedData<U> pagedData, @Nonnull BeanTransformer<U, V> transformer) {
+    public static <U extends Bean, V extends Bean> PagedData<V> transform(
+            @Nonnull PagedData<U> pagedData, @Nonnull BeanTransformer<U, V> transformer
+    ) {
         PagedData<V> p = new PagedData<>();
         p.setCount(pagedData.getCount());
         p.setCurrentPage(pagedData.getCurrentPage());
@@ -205,8 +233,9 @@ public final class PagingUtil {
      * @return 第一个类型的PagedData。
      */
     @SuppressWarnings("DuplicatedCode")
-    public static <U extends Bean, V extends Bean> PagedData<U>
-    reverseTransform(@Nonnull PagedData<V> pagedData, @Nonnull BeanTransformer<U, V> transformer) {
+    public static <U extends Bean, V extends Bean> PagedData<U> reverseTransform(
+            @Nonnull PagedData<V> pagedData, @Nonnull BeanTransformer<U, V> transformer
+    ) {
         PagedData<U> p = new PagedData<>();
         p.setCount(pagedData.getCount());
         p.setCurrentPage(pagedData.getCurrentPage());
@@ -219,17 +248,31 @@ public final class PagingUtil {
     /**
      * 返回指定的列表在指定的分页上对应的子列表。
      *
+     * <p>
+     * 必须保证 <code>pagingInfo</code> 参数的页数和行数是合法的，即
+     * <code>page &gt;= 0</code> 且 <code>rows &gt;= 0</code>，
+     * 否则会抛出异常。
+     *
      * @param list       指定的列表。
      * @param pagingInfo 指定的分页。
      * @param <E>        列表中的元素类型。
      * @return 指定的列表在指定的分页上对应的子列表。
+     * @throws IllegalArgumentException 参数不合法。
      */
     public static <E> List<E> subList(List<E> list, PagingInfo pagingInfo) {
-        int beginIndex = pagingInfo.getRows() * pagingInfo.getPage();
+        int page = pagingInfo.getPage();
+        int rows = pagingInfo.getRows();
+        if (page < 0 || rows < 0) {
+            throw new IllegalArgumentException("分页信息的页数必须大于等于 0, 行数必须大于等于 0");
+        }
+        if (rows == 0) {
+            return Collections.emptyList();
+        }
+        int beginIndex = rows * page;
         if (beginIndex >= list.size()) {
             return Collections.emptyList();
         }
-        int endIndex = Math.min(list.size(), beginIndex + pagingInfo.getRows());
+        int endIndex = Math.min(list.size(), beginIndex + rows);
         return list.subList(beginIndex, endIndex);
     }
 }