+++ subgrade-stack/src/main/java/com/dwarfeng/subgrade/stack/bean/dto/PagedData.java
@@ -13,21 +13,29 @@ public class PagedData<E> implements Dto {
     private static final long serialVersionUID = 9121659227089130120L;
 
     /**
-     * 当前的页数，从0开始计数。
+     * 当前的页数，从 0 开始计数。
      */
     private int currentPage;
+
     /**
-     * 总共的页数，该值最小是1。
+     * 总共的页数。
+     *
+     * <p>
+     * 在一般情况下，该值是一个非负数，代表总共的页数。<br>
+     * 特殊情况下，该值是负数，代表总页数无法计算，这通常在分页查询时，将每页返回的行数设置为 0 时出现。
      */
     private int totalPages;
+
     /**
      * 每页返回的行数。
      */
     private int rows;
+
     /**
      * 总共数量。
      */
     private long count;
+
     /**
      * 当前页数据。
      */